CIDR

Description

Experimenting with programming in CIDR shallow embedding in Haskell/AERN2.

CIDR is a core language for exact real computation being developed within the CID EU project.

Documentation

sqrtApprox :: _ => Rational -> r -> EnsureCN r

sqrtApprox k x

preconditions: x > 1, k >= 0

returns a real number that is within $2^{{}-k}$ away from $sqrt{x}$.

data UseLipschitz

Constructors

UseLipschitz_YES

UseLipschitz_NO

mysqrt :: _ => r -> EnsureCN r

mysqrtX :: _ => UseLipschitz -> r -> EnsureCN r

mysqrt useLipschitz useNearbySimpler x

preconditions: x > 1

returns $sqrt{x}$.

mysqrtNx :: _ => UseLipschitz -> Integer -> r -> r

A n-times nested sqrt:

mysqrtNx_CR_test :: CauchyReal

mysqrtNx_vanilla_test :: MPBall

mysqrtNx_Lip_test :: MPBall

class CanMVApproxCompare t1 t2 where

Minimal complete definition

mvApproxEq

Associated Types

type CanMVApproxCompareType t1 t2

Methods

mvApproxEq :: Rational -> t1 -> t2 -> CanMVApproxCompareType t1 t2

mvIsPositiveUpTo p x

Return true if x is above $2^{{}-p}$.

Return false if the number is negative.

Return true or false if the number is between 0 and $2^{{}-p}$.

while :: (CanChoose b (EnsureCN t), HasBools b, CanEnsureCN t, Show t) => t -> (t -> b) -> (t -> t) -> EnsureCN t

class CanEnsureCN t => CanChoose b t where

Minimal complete definition

choose

Methods

choose :: [(b, t)] -> EnsureCN t

Return any of the values in the associated list whose key is true. This is useful when the conditions are only semi-decidable. In this case the conditions should be evaluated in a parallel or interleaved way so that when some tests do not terminate and others do, the whole expression terminates.

class CanFindNearbySimpler r where

Minimal complete definition

nearbySimpler

Methods

nearbySimpler :: Rational -> r -> r