Tasks.Fourier

Documentation

taskFFTDescription :: Integer -> Integer -> Integer -> String

taskFFT :: (FFTOps c, HasIntegers c) => (CauchyReal -> c) -> Integer -> [c]

taskFFTWithHook :: (FFTOps c, HasIntegers c) => (CauchyReal -> c) -> (c -> Maybe c) -> Integer -> Maybe [c]

taskFFTA :: (FFTOpsA r c, QAArrow to, HasIntegers (QA to p), c ~ Complex (QA to p), QAProtocolCacheable p, ConvertibleExactly CauchyReal (QA to p)) => Integer -> () `to` [c]

type FFTOpsA r c = (CanAddSubMulBy r r, c ~ Complex r)

type FFTOps c = (CanMulBy c Rational, CanMulBy c (Complex Integer), CanAddSubMulBy c c, CanNegSameType c, CanExpSameType c)

ditfft2 :: FFTOps c => (CauchyReal -> c) -> Integer -> [c] -> [c]

Cooley-Tukey FFT closely following https://en.wikipedia.org/wiki/Cooley%E2%80%93Tukey_FFT_algorithm#Pseudocode

A vanilla version using Data.Sequence.

ditfft2Hook :: FFTOps c => (CauchyReal -> c) -> (c -> Maybe c) -> Integer -> [c] -> Maybe [c]

Cooley-Tukey FFT closely following https://en.wikipedia.org/wiki/Cooley%E2%80%93Tukey_FFT_algorithm#Pseudocode

A version with a hook that allows one to stop the computation at any point and propagate Nothing.

ditfft2A :: (FFTOpsA r c, QAArrow to, c ~ Complex (QA to p), QAProtocolCacheable p, ConvertibleExactly CauchyReal (QA to p)) => Integer -> [c] `to` [c]

Cooley-Tukey FFT closely following https://en.wikipedia.org/wiki/Cooley%E2%80%93Tukey_FFT_algorithm#Pseudocode

regComplex :: (QAArrow to, QAProtocolCacheable p) => Bool -> Complex (QA to p) `to` Complex (QA to p)

complexGiveName :: String -> Complex (QA to p) -> Complex (QA to p)

twsCR :: Integer -> Integer -> Map Rational (Complex CauchyReal)

tws :: FFTOps c => (CauchyReal -> c) -> Integer -> Integer -> Map Rational c

taskDFTDescription :: Integer -> Integer -> Integer -> String

taskDFT :: (FFTOps c, HasIntegers c) => (CauchyReal -> c) -> Integer -> [c]

taskDFTWithHook :: (FFTOps c, HasIntegers c) => (CauchyReal -> c) -> (c -> Maybe c) -> Integer -> Maybe [c]

dft :: (FFTOps c, HasIntegers c) => (CauchyReal -> c) -> (c -> Maybe c) -> [c] -> Maybe [c]

dftA :: (FFTOpsA r c, QAArrow to, HasIntegers c, CanAddSubMulBy r (CauchyRealA to)) => (String -> c `to` Maybe c) -> [c] `to` Maybe [c]